backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
ortools>=9.8.3296
pandas>=2.2.0
openpyxl>=3.1.2
pydantic>=2.6.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
